# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API CRUD Mahasiswa dan Jurusan",
        "title": "Kampus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/jurusan": {
            "get": {
                "description": "Menampilkan seluruh data jurusan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jurusan"
                ],
                "summary": "Ambil semua jurusan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Jurusan"
                            }
                        }
                    }
                }
            }
        },
        "/mahasiswa": {
            "get": {
                "description": "Menampilkan seluruh data mahasiswa",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil semua mahasiswa",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Mahasiswa"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan data mahasiswa baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah mahasiswa",
                "parameters": [
                    {
                        "description": "Data Mahasiswa",
                        "name": "mahasiswa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Mahasiswa"
                        }
                    }
                }
            }
        },
        "/mahasiswa/{id}": {
            "put": {
                "description": "Mengubah data mahasiswa berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Update mahasiswa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasiswa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Mahasiswa",
                        "name": "mahasiswa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Mahasiswa"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data mahasiswa berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus mahasiswa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasiswa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Mencari mahasiswa berdasarkan nama",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Cari mahasiswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Mahasiswa",
                        "name": "search",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Mahasiswa"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Jurusan": {
            "type": "object",
            "properties": {
                "fakultas": {
                    "type": "string"
                },
                "idjurusan": {
                    "type": "integer"
                },
                "jenjang": {
                    "type": "string"
                },
                "namaJurusan": {
                    "type": "string"
                }
            }
        },
        "models.Mahasiswa": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idjurusan": {
                    "type": "integer"
                },
                "jurusan": {
                    "$ref": "#/definitions/models.Jurusan"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "tglLahir": {
                    "type": "string"
                },
                "umur": {
                    "type": "integer"
                }
            }
        }
    }
}